\name{print.netgwas}
\alias{print.netgwas}
\title{
Print function for S3 class "netgwas"
}
\description{
Print a summary of results from function \code{\link{netsnp}}, \code{\link{netphenogeno}}. 
}
\usage{
\method{print}{netgwas}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class "netgwas"
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{ 
  Pariya Behrouzi and Ernst C. Wit \cr
  Maintainer: Pariya Behrouzi \email{pariya.behrouzi@gmail.com} \cr
}

\references{
Behrouzi, P., and Wit, E. C. (2017c). netgwas: An R Package for Network-Based Genome-Wide Association Studies. arXiv preprint, arXiv:1710.01236.
}

\seealso{
\code{\link{netmap}}, \code{\link{netsnp}}, \code{\link{netphenogeno}}
}