\name{print.netgwasmap}
\alias{print.netgwasmap}
\title{
Print function for S3 class "netgwasmap"
}
\description{
Print a summary of results from function \code{\link{netmap}}. 
}
\usage{
\method{print}{netgwasmap}(x, ...)
}
\arguments{
  \item{x}{
An object with S3 class "netgwasmap"
}
  \item{\dots}{
System reserved (No specific usage)
}
}

\author{ 
  Pariya Behrouzi and Ernst C. Wit \cr
  Maintainer: Pariya Behrouzi \email{pariya.behrouzi@gmail.com} \cr
}

\references{
Behrouzi, P., and Wit, E. C. (2017c). netgwas: An R Package for Network-Based Genome-Wide Association Studies. arXiv preprint, arXiv:1710.01236.
}

\seealso{
\code{\link{netmap}}
}